+++ server/src/main/java/net/feedbacky/app/StartupMigrator.java
@@ -10,12 +10,15 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.PersistenceContextType;
+import javax.persistence.Query;
 
+import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileNotFoundException;
+import java.io.FileReader;
 import java.io.IOException;
-import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -29,9 +32,11 @@ import java.util.logging.Logger;
 @Order
 public class StartupMigrator {
 
-  public static final int FILE_VERSION = 2;
+  public static final int FILE_VERSION = 3;
   private final Logger logger = Logger.getLogger("Migrator");
   private final UserRepository userRepository;
+  @PersistenceContext(type = PersistenceContextType.EXTENDED)
+  private EntityManager entityManager;
   private int version = -1;
   private File versionFile;
 
@@ -51,8 +56,14 @@ public class StartupMigrator {
         doAttemptMigration();
         return;
       }
-      String content = new String(Files.readAllBytes(versionFile.toPath()), StandardCharsets.UTF_8);
-      version = parseIntOrDefault(content, FILE_VERSION);
+      try(BufferedReader reader = new BufferedReader(new FileReader(versionFile))) {
+        String content = reader.readLine();
+        if(content == null || content.equals("")) {
+          version = FILE_VERSION;
+        } else {
+          version = Integer.parseInt(content);
+        }
+      }
     } catch(IOException e) {
       logger.log(Level.WARNING, "Failed to get version file contents! Migrator will not run.");
       e.printStackTrace();
@@ -66,22 +77,11 @@ public class StartupMigrator {
   private void doAttemptMigration() {
     for(int i = version; i < FILE_VERSION; i++) {
       switch(i) {
+        case 2:
+          improvedMailNotificationsFeatureMigration();
+          break;
         case 1:
-          logger.log(Level.INFO, "Migrating Feedbacky from version 1 to 2...");
-          for(User user : userRepository.findAll()) {
-            if(user.getMailPreferences() != null) {
-              continue;
-            }
-            MailPreferences defaultPreferences = new MailPreferences();
-            defaultPreferences.setNotifyFromModeratorsComments(true);
-            defaultPreferences.setNotifyFromStatusChange(true);
-            defaultPreferences.setNotifyFromTagsChange(true);
-            defaultPreferences.setUnsubscribeToken(RandomStringUtils.randomAlphanumeric(6));
-            defaultPreferences.setUser(user);
-            user.setMailPreferences(defaultPreferences);
-            userRepository.save(user);
-          }
-          logger.log(Level.INFO, "Migrated to version 2.");
+          mailPreferencesFeatureMigration();
           break;
         case -1:
           break;
@@ -90,22 +90,77 @@ public class StartupMigrator {
     saveMigrationFile();
   }
 
-  private void saveMigrationFile() {
-    try(PrintStream out = new PrintStream(versionFile)) {
-      out.print(FILE_VERSION + "");
-    } catch(FileNotFoundException e) {
-      e.printStackTrace();
+  private void improvedMailNotificationsFeatureMigration() {
+    logger.log(Level.INFO, "Migrating Feedbacky from version 2 to 3...");
+    logger.log(Level.INFO, "It may take some time depending on users amount in database.");
+    int affected = 0;
+    boolean missingValues = false;
+    for(User user : userRepository.findAll()) {
+      int notificationsEnabledAmount = 0;
+      MailPreferences preferences = user.getMailPreferences();
+      if(missingValues) {
+        preferences.setNotificationsEnabled(true);
+        user.setMailPreferences(preferences);
+        userRepository.save(user);
+        affected++;
+        continue;
+      }
+      Query query = entityManager.createNativeQuery("SELECT notify_from_moderators_comments, notify_from_status_change, notify_from_tags_change FROM users_mail_preferences WHERE user_id = ?");
+      query.setParameter(1, user.getId());
+      Object[] result;
+      try {
+        result = (Query[]) query.getSingleResult();
+      } catch(Exception e) {
+        logger.log(Level.WARNING, "Encountered SQL exception, probably missing some fields to migrate, switching to non SQL migration.");
+        missingValues = true;
+        preferences.setNotificationsEnabled(true);
+        user.setMailPreferences(preferences);
+        userRepository.save(user);
+        affected++;
+        continue;
+      }
+      if((boolean) result[0]) {
+        notificationsEnabledAmount++;
+      }
+      if((boolean) result[1]) {
+        notificationsEnabledAmount++;
+      }
+      if((boolean) result[2]) {
+        notificationsEnabledAmount++;
+      }
+      //if 2 out of 3 notification methods are enabled assume that user want to receive future notifications about all types of events
+      if(notificationsEnabledAmount >= 2) {
+        preferences.setNotificationsEnabled(true);
+        user.setMailPreferences(preferences);
+        userRepository.save(user);
+        affected++;
+      }
     }
+    logger.log(Level.INFO, "Migrated to version 3, affected entities: {0}.", affected);
   }
 
-  private int parseIntOrDefault(String string, int defaultValue) {
-    if(string.equals("")) {
-      return defaultValue;
+  private void mailPreferencesFeatureMigration() {
+    logger.log(Level.INFO, "Migrating Feedbacky from version 1 to 2...");
+    logger.log(Level.INFO, "It may take some time depending on users amount in database.");
+    for(User user : userRepository.findAll()) {
+      if(user.getMailPreferences() != null) {
+        continue;
+      }
+      MailPreferences defaultPreferences = new MailPreferences();
+      defaultPreferences.setNotificationsEnabled(true);
+      defaultPreferences.setUnsubscribeToken(RandomStringUtils.randomAlphanumeric(6));
+      defaultPreferences.setUser(user);
+      user.setMailPreferences(defaultPreferences);
+      userRepository.save(user);
     }
+    logger.log(Level.INFO, "Migrated to version 2.");
+  }
+
+  private void saveMigrationFile() {
     try {
-      return Integer.parseInt(string);
-    } catch(Exception ex) {
-      return defaultValue;
+      Files.write(versionFile.toPath(), String.valueOf(FILE_VERSION).getBytes());
+    } catch(IOException e) {
+      e.printStackTrace();
     }
   }