+++ server/src/main/java/net/feedbacky/app/oauth/provider/google/GoogleLoginController.java
@@ -136,9 +136,7 @@ public class GoogleLoginController implements AbstractLoginProvider {
       user.setUsername(googleUser.getName());
       MailPreferences preferences = new MailPreferences();
       preferences.setUnsubscribeToken(RandomStringUtils.randomAlphanumeric(6));
-      preferences.setNotifyFromTagsChange(true);
-      preferences.setNotifyFromStatusChange(true);
-      preferences.setNotifyFromModeratorsComments(true);
+      preferences.setNotificationsEnabled(true);
       preferences.setUser(user);
       user.setMailPreferences(preferences);
       Set<ConnectedAccount> accounts = new HashSet<>(user.getConnectedAccounts());